+++ test/component.spec.js
@@ -17,6 +17,7 @@ import TransWithNumberParams from './components/TransWithNumberParams'
 import TransWithJunkParams from './components/TransWithJunkParams'
 import Dates from './components/Dates'
 import {TransPluralize1, TransPluralize2} from './components/TransPlurals'
+import TransWithObjParams from './components/TransWithObjParams'
 
 const translations = {
   'es': {
@@ -105,6 +106,14 @@ describe('component test', function() {
       </Provider>
     ))
 
+    this.objAsParam = ReactDOM.findDOMNode(TestUtils.renderIntoDocument(
+      <Provider store={this.store}>
+        <I18n translations={translations}>
+          <TransWithObjParams/>
+        </I18n>
+      </Provider>
+    ))
+
   })
 
   it('initial state', function() {
@@ -169,6 +178,10 @@ describe('component test', function() {
     this.store.dispatch(setLanguage('es-ES'))
     expect(this.store.getState().i18nState.lang).toEqual('es-ES')
     expect(this.withoutParamsNode.textContent).toEqual('Hola')
+  })
+
+  it('object as param', function() {
+    expect(this.objAsParam.textContent).toEqual('Hello <span>Cesc</span>')
   })
 
 })