+++ test/components/TransWithObjParams.js
@@ -1,0 +1,13 @@
+import React from 'react'
+
+const TransWithObjParams = ({}, context) => {
+  const user = {name: 'Cesc'}
+  const name = <span>{user.name}</span>
+  return <div>{context.t('Hello {name}', {name: name})}</div>
+}
+
+TransWithObjParams.contextTypes = {
+  t: React.PropTypes.func.isRequired
+}
+
+export default TransWithObjParams