+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 1.1.5 (Not created yet)
 - Changelog format improved.
 - Documentation improved.
+- Now is possible use html object as parameter.
 
 ## 1.1.4
 - Fall back lang. Read documentation https://github.com/APSL/redux-i18n#changing-language